Potential fixes to #1 and #5 & updated ServiceStack
Fixing #5: Importing database crash

This was due to multiple threads trying to access sqlite at once.
Cacher now locks to prevent this and is also transactionalised.
Due to some very odd behaviour from either ormlite or sqlite,
I've had to transactionalise EVERYTHING otherwise any operation
that isn't transactionalised fails.

Fixing #1: Names being blank

I think this was down to the code only using the card ID when
no user was found in the db; it should also have done this when
the user was present but had no name set (the new behaviour).