Bugfix(QuiRing): BeginInvoke calls missing from some handlers
Some of the handlers in QuiRing for events fired up from Quiche
were not using if(!InvokeRequired).... else(BeginInvoke)
and so were causing threading errors on the GUI thread. They
should now be correctly BeginInvoked as appropriate.

Also CardsTab.OnCardRead took Zones and Terminal arguments the
wrong way round compared to the order expected everywhere else.
That has been switched to agree with the rest of the code.